--- src/components/atoms/MobileAgendaView/MobileAgendaView.types.ts
@@ -1,5 +1,0 @@
-import { Event } from '@atoms/EventCalendar/EventCalendar.types'
-
-export type MobileAgendaViewProps = {
-  events: Event[]
-}
