+++ src/components/atoms/PageTabs/PageTabs.tsx
@@ -3,6 +3,12 @@ import { PageTabsProps } from './PageTabs.types'
 
 /** A component displaying tab buttons that can be clicked to change the active tab.
  * Used within page content (below main navigation) as a sub-navigation system.
+ * Example:
+ *       const [activeTab, setActiveTab] = useState<string>('News')
+ *       ...
+ *       <ContentBlock>
+ *         <PageTabs tabs={['News', 'Events']} activeTab={activeTab} setActiveTab={setActiveTab} />
+ *       </ContentBlock>
  */
 const PageTabs = (props: PageTabsProps) => {
   const { tabs, activeTab, setActiveTab } = props
