+++ src/components/molecules/MobileAgendaView/MobileAgendaView.types.ts
@@ -1,0 +1,5 @@
+import { Event } from '@organisms/EventCalendar/EventCalendar.types'
+
+export type MobileAgendaViewProps = {
+  events: Event[]
+}
