+++ src/pages/events/index.tsx
@@ -1,5 +1,5 @@
 import { ContentBlock } from '@atoms/ContentBlock'
-import { EventCalendar } from '@atoms/EventCalendar'
+import { EventCalendar } from '@organisms/EventCalendar'
 
 /** TODO: Tear down once events are fully finished */
 const EventsPage = () => {
