+++ src/pages/news/post/[id]/index.tsx
