+++ tsconfig.json
@@ -28,6 +28,6 @@
       "@models/*": ["src/models/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "src/components/atoms/EventCalendar/EventCalendar.types.ts"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "src/components/organisms/EventCalendar/EventCalendar.types.ts"],
   "exclude": ["node_modules"]
 }
